Patient infection route in Vietnam

This dataset contained information of patients linkage as well as other information of the infected. This list is not exhaustive as not all infected sources can be identified. Information is gathered based on the MOH's daily announcement of patients' travelling history which was last updated by 10 Oct 2020.

patient_link

Format

An object of class data.frame with 975 rows and 17 columns.

Source

Vietnam's MOH at https://ncov.moh.gov.vn

Details

The dataset contained the following variables:

from:	The patient id of source of infection

to:	The patient id of the infected

dob:	The date of birth of the infected

age:	The age of the infected by 2020

gender:	The gender of the infected

address:	living address of the infected

treatment_location:	the treatment location of the infected. Could differ from the address as patients are treated focusly in some specialised medical centres.

status:	either "Active", "Recovered", "Deceased" or "Other". Other indicated patients'death of other causes rather than of Covid-19.

type:	type of infection. Either "foreign expert", from oversea", "health-care worker", "illegal immigrated", "quarantine upon entry" or "social setting'. Meaning of each type is noted in the app

nationality:	nationality of the infected

date_announced:	date announced of the confirmed case

month_announced:	month announced of the confirmed case

death_announced:	date announced of the deceased

vessel_no:	The vessel that transported the infected into Vietnam; for cases coming back from oversea only.

flight_from:	The departed destination of the infected before coming into Vietnam; for cases coming back from oversea only.

note:	Additional note