Return list of references used in the app

the ref function returns the list of R packages and external resources refered in the app. As both packages and external links are gathered into one data frame, duplication of values may occur.

ref

Format

An object of class data.frame with 12 rows and 2 columns.

Examples

ref

#>              packages
#> 1               shiny
#> 2      shinydashboard
#> 3  shinydashboardPlus
#> 4           tidyverse
#> 5              plotly
#> 6           lubridate
#> 7             leaflet
#> 8               rgdal
#> 9              tigris
#> 10          geojsonio
#> 11                 DT
#> 12          echarts4r
#>                                                                                                                                                                                                                                                link
#> 1                                                                                                                                                       Wei, S. (2020).Covid-2019.live. Retrieved 9 October 2020, from https://covid-2019.live/en/.
#> 2                                                                                                                                         Face Mask Man icon. UXWing. (2020). Retrieved 9 October 2020, from https://uxwing.com/face-mask-man-icon/
#> 3                                                                  Le, B. (2020). Vietnam Choropleth map. Raw.githubusercontent.com. Retrieved 9 October 2020, from https://raw.githubusercontent.com/lebinh/vietnam-choropleth-map/master/vn.json.
#> 4    Sievert, C. (2020). 17 Server-side linking with shiny | Interactive web-based data visualization with R, plotly, and shiny. Plotly-r.com. Retrieved 9 October 2020, from https://plotly-r.com/linking-views-with-shiny.html#shiny-performance.
#> 5  So ca nhiem COVID19 theo tinh thanh Viet Nam. Vi.wikipedia.org. (2020). Retrieved 9 October 2020, from https://vi.wikipedia.org/wiki/B%E1%BA%A3n_m%E1%BA%ABu:S%E1%BB%91_ca_nhi%E1%BB%85m_COVID-19_theo_t%E1%BB%89nh_th%C3%A0nh_Vi%E1%BB%87t_Nam.
#> 6                                                                                                TRANG TIN VE DICH BENH VIEM DUONG HO HAP CAP COVID-19 - BO Y Te(2020). Retrieved 9 October 2020, from https://ncov.moh.gov.vn/web/guest/trang-chu.
#> 7                                                                                                                                                       Wei, S. (2020).Covid-2019.live. Retrieved 9 October 2020, from https://covid-2019.live/en/.
#> 8                                                                                                                                         Face Mask Man icon. UXWing. (2020). Retrieved 9 October 2020, from https://uxwing.com/face-mask-man-icon/
#> 9                                                                  Le, B. (2020). Vietnam Choropleth map. Raw.githubusercontent.com. Retrieved 9 October 2020, from https://raw.githubusercontent.com/lebinh/vietnam-choropleth-map/master/vn.json.
#> 10   Sievert, C. (2020). 17 Server-side linking with shiny | Interactive web-based data visualization with R, plotly, and shiny. Plotly-r.com. Retrieved 9 October 2020, from https://plotly-r.com/linking-views-with-shiny.html#shiny-performance.
#> 11 So ca nhiem COVID19 theo tinh thanh Viet Nam. Vi.wikipedia.org. (2020). Retrieved 9 October 2020, from https://vi.wikipedia.org/wiki/B%E1%BA%A3n_m%E1%BA%ABu:S%E1%BB%91_ca_nhi%E1%BB%85m_COVID-19_theo_t%E1%BB%89nh_th%C3%A0nh_Vi%E1%BB%87t_Nam.
#> 12                                                                                               TRANG TIN VE DICH BENH VIEM DUONG HO HAP CAP COVID-19 - BO Y Te(2020). Retrieved 9 October 2020, from https://ncov.moh.gov.vn/web/guest/trang-chu.